requirements.txt
flask
flask_login
psycopg[binary]
gunicorn
Flask-SQLAlchemy
Flask-WTF
python-dotenv
flask-ckeditor
